+++ bcbio/install.py
@@ -198,12 +198,16 @@ def _get_conda_bin():
         return conda_bin
 
 def _default_deploy_args(args):
+    """Standard install arguments for CloudBioLinux.
+
+    Avoid using sudo and keep an installation isolated if running as the root user.
+    """
     return {"flavor": "ngs_pipeline_minimal",
             "vm_provider": "novm",
             "hostname": "localhost",
             "fabricrc_overrides": {"edition": "minimal",
                                    "use_sudo": False,
-                                   "keep_isolated": args.isolate,
+                                   "keep_isolated": args.isolate or os.geteuid() == 0,
                                    "conda_cmd": _get_conda_bin(),
                                    "distribution": args.distribution or "__auto__",
                                    "dist_name": "__auto__"}}
@@ -219,7 +223,7 @@ def _update_conda_packages():
         os.remove(req_file)
     subprocess.check_call(["wget", "-O", req_file, "--no-check-certificate", REMOTES["requirements"]])
     subprocess.check_call([conda_bin, "install", "--update-deps", "--quiet", "--yes",
-                           "-c", "bioconda", "--file", req_file])
+                           "-c", "bioconda", "-c", "conda-forge", "--file", req_file])
     if os.path.exists(req_file):
         os.remove(req_file)
     return os.path.dirname(os.path.dirname(conda_bin))
@@ -230,7 +234,7 @@ def _update_conda_devel():
     conda_bin = _get_conda_bin()
     assert conda_bin, "Could not find anaconda distribution for upgrading bcbio"
     subprocess.check_call([conda_bin, "install", "--update-deps",
-                           "--quiet", "--yes", "-c", "bioconda", "bcbio-nextgen"])
+                           "--quiet", "--yes", "-c", "bioconda", "-c", "conda-forge", "bcbio-nextgen"])
     return os.path.dirname(os.path.dirname(conda_bin))
 
 def get_genome_dir(gid, galaxy_dir, data):