+++ bcbio/pipeline/sample.py
@@ -115,8 +115,7 @@ def process_alignment(data, alt_input=None):
     if fastq1 and objectstore.file_exists_or_remote(fastq1) and aligner:
         logger.info("Aligning lane %s with %s aligner" % (data["rgnames"]["lane"], aligner))
         data = align_to_sort_bam(fastq1, fastq2, aligner, data)
-        umi_file = dd.get_umi_file(data)
-        if umi_file:
+        if dd.get_umi_consensus(data):
             data["umi_bam"] = dd.get_work_bam(data)
             if fastq2:
                 f1, f2 = postalign.umi_consensus(data)