+++ tests/conftest.py
@@ -82,7 +82,7 @@ def install_test_files(data_dir):
         DlInfo("110106_FC70BUKAAXX.tar.gz", None, None),
         DlInfo("genomes_automated_test.tar.gz", "genomes", 31),
         DlInfo("110907_ERP000591.tar.gz", None, None),
-        DlInfo("100326_FC6107FAAXX.tar.gz", None, 10),
+        DlInfo("100326_FC6107FAAXX.tar.gz", None, 11),
         DlInfo("tcga_benchmark.tar.gz", None, 3),
         DlInfo("singlecell-rnaseq-test-data.tar.gz", "Harvard-inDrop", 1)
     ]