+++ HISTORY.md
@@ -12,6 +12,8 @@
   https://github.com/bcbio/bcbio.github.io/blob/master/_posts/2016-11-21-giab-hg38-freebayes.md
 - Fixes for Seq2C concurrency problems when preparing BED files. Thanks to Vlad
   Saveliev.
+- Pre-install capture region BED files and enable short hand specification in
+  sample configuration.
 - Testing suite cleanup. Move to py.test and separate integration and unit
   tests. Thanks to Tetiana Khotiainsteva.